added demos
